requirements.txt
crewai==0.41.1
crewai-tools==0.4.26
panel==1.3.8
streamlit==1.28.0
requests==2.32.4
pyyaml==6.0.1
pygithub==2.1.1
google-generativeai==0.3.2
python-dotenv==1.0.0